+++ tests/test_csc.py
@@ -354,6 +354,10 @@ class TestRotatorCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCas
             self.enable_mock_ccw_telemetry = False
             await asyncio.wait_for(self.mock_ccw_task, timeout=STD_TIMEOUT)
             await self.assert_next_summary_state(salobj.State.FAULT)
+
+            # TODO: At ts_xml v23.0.0, use ErrorCode.NO_NEW_CCW_TELEMETRY
+            # instead.
+            await self.assert_next_sample(topic=self.remote.evt_errorCode, errorCode=4)
             await self.assert_next_sample(
                 topic=self.remote.evt_errorCode, errorCode=ErrorCode.CONTROLLER_FAULT
             )
@@ -388,6 +392,10 @@ class TestRotatorCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCas
             self.ccw_following_error = self.csc.config.max_ccw_following_error + 0.1
             await self.assert_next_ccw_following_error()
             await self.assert_next_summary_state(salobj.State.FAULT)
+
+            # TODO: At ts_xml v23.0.0, use ErrorCode.CCW_FOLLOWING_ERROR
+            # instead.
+            await self.assert_next_sample(topic=self.remote.evt_errorCode, errorCode=6)
             await self.assert_next_sample(
                 topic=self.remote.evt_errorCode, errorCode=ErrorCode.CONTROLLER_FAULT
             )
@@ -412,6 +420,10 @@ class TestRotatorCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCas
             self.ccw_following_error = -(self.csc.config.max_ccw_following_error + 0.1)
             await self.assert_next_ccw_following_error()
             await self.assert_next_summary_state(salobj.State.FAULT)
+
+            # TODO: At ts_xml v23.0.0, use ErrorCode.CCW_FOLLOWING_ERROR
+            # instead.
+            await self.assert_next_sample(topic=self.remote.evt_errorCode, errorCode=6)
             await self.assert_next_sample(
                 topic=self.remote.evt_errorCode,
                 errorCode=ErrorCode.CONTROLLER_FAULT,
@@ -444,6 +456,10 @@ class TestRotatorCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCas
             await self.assert_next_summary_state(
                 salobj.State.FAULT, flush=False, timeout=STD_TIMEOUT
             )
+
+            # TODO: At ts_xml v23.0.0, use ErrorCode.CCW_FOLLOWING_ERROR
+            # instead.
+            await self.assert_next_sample(topic=self.remote.evt_errorCode, errorCode=6)
             await self.assert_next_sample(
                 topic=self.remote.evt_errorCode,
                 errorCode=ErrorCode.CONTROLLER_FAULT,
@@ -989,6 +1005,9 @@ class TestRotatorCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCas
                 flush=False, timeout=STD_TIMEOUT
             )
             self.assertFalse(data.inPosition)
+
+            self.remote.evt_errorCode.flush()
+
             await self.remote.cmd_trackStart.start(timeout=STD_TIMEOUT)
 
             # Immediately send a track commands (not giving the CSC time to see
@@ -1022,6 +1041,10 @@ class TestRotatorCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCas
             await self.assert_next_sample(
                 topic=self.remote.evt_tracking, tracking=False, noNewCommand=True
             )
+
+            # TODO: At ts_xml v23.0.0, use ErrorCode.NO_NEW_TRACK_COMMAND
+            # instead.
+            await self.assert_next_sample(topic=self.remote.evt_errorCode, errorCode=7)
 
     async def assert_next_ccw_following_error(
         self,